+++ examples/vike-react-simple/pages/+config.js
@@ -1,5 +1,0 @@
-import vikeReactSimple from 'vike-react-simple/config'
-
-export default {
-  extends: [vikeReactSimple]
-}