+++ examples/webpack/vite.config.js
@@ -1,5 +1,0 @@
-import vike from 'vike/plugin'
-
-export default {
-  plugins: [vike()]
-}