+++ test/misc/pages/about/Page.jsx
@@ -1,3 +1,5 @@
+// Test: define `Page` over `+config.js > export default { Page }` instead of +Page.js
+
 export default Page
 
 import React from 'react'