+++ vike/node/plugin/plugins/importUserCode/v1-design/getVikeConfig/configDefinitionsBuiltIn.ts
@@ -117,7 +117,7 @@ const configDefinitionsBuiltIn: ConfigDefinitionsBuiltIn = {
     env: { server: true, client: true }
   },
   passToClient: {
-    env: { server: true, config: true },
+    env: { server: true },
     cumulative: true
   },
   route: {
@@ -165,7 +165,11 @@ const configDefinitionsBuiltIn: ConfigDefinitionsBuiltIn = {
   meta: {
     env: { config: true }
   },
-  clientEntryLoaded: {
+  // Whether the page loads:
+  //  - Vike's client runtime
+  //  - User's client hooks
+  // In other words, whether the page is "HTML-only" (https://vike.dev/render-modes). HTML-only pages shouldn't load the client runtime nor client hooks.
+  isClientRuntimeLoaded: {
     env: { server: true, client: true },
     eager: true,
     _computed: (configValueSources): boolean => {