+++ vike/node/plugin/plugins/importUserCode/v1-design/getVikeConfig/transpileAndExecuteFile.ts
@@ -235,11 +235,10 @@ async function transpileWithEsbuild(
           if (debug.isActivated)
             debug('onResolved()', { args, resolved, importPathTranspiled, isPointerImport, isExternal })
 
-          pointerImports[importPathTranspiled] = isPointerImport
           if (isExternal) {
+            pointerImports[importPathTranspiled] = isPointerImport
             return { external: true, path: importPathTranspiled }
           } else {
-            resolved.path = importPathTranspiled
             return resolved
           }
         })